+++ tests/Integration/Get-DscSplattedResource.Integration.Tests.ps1
@@ -5,9 +5,10 @@ BeforeDiscovery {
     $env:PSModulePath = "$modulePath;$($env:PSModulePath)"
     $dscResources = Get-DscResource -Name MofBased*, ClassBased* -ErrorAction SilentlyContinue
 
-    $skippedDscResources = 'ClassBasedResource3', 'MofBasedResource2', 'MofBasedResource3'
+    $skippedDscResources = ''
 
     Import-Module -Name datum
+    Import-Module -Name DscBuildHelpers -Force
 
     $datum = New-DatumStructure -DefinitionFile $here\Assets\Datum.yml
     $allNodes = Get-Content -Path $here\Assets\AllNodes.yml -Raw | ConvertFrom-Yaml